refactor: use value selector for pause timeout instead of time stepper
Changed the pause timeout setting from a time stepper to a value selector with predefined options (30s, 1m, 2m, 3m, 4m, 5m). This provides a better UX by offering clear, discrete choices rather than a continuous range.

Changes:
- Updated PauseSettingsScreenModel to use Int instead of Long for timeout
- Replaced PreferenceTimeStepper with PreferenceValueSelector
- Defined six timeout values: 30s, 1m, 2m, 3m, 4m, 5m
- Added button and display formatters for user-friendly labels

+++ app/src/main/java/com/enaboapps/switchify/screens/settings/pause/models/PauseSettingsScreenModel.kt
@@ -18,14 +18,14 @@ class PauseSettingsScreenModel(context: Context) : ViewModel() {
     }
     val holdToUnpauseDuration: LiveData<Long> = _holdToUnpauseDuration
 
-    private val _pauseTimeout = MutableLiveData<Long>().apply {
+    private val _pauseTimeout = MutableLiveData<Int>().apply {
         value =
             preferenceManager.getLongValue(
                 PreferenceManager.Keys.PREFERENCE_KEY_PAUSE_TIMEOUT,
                 30000L
-            )
+            ).toInt()
     }
-    val pauseTimeout: LiveData<Long> = _pauseTimeout
+    val pauseTimeout: LiveData<Int> = _pauseTimeout
 
     fun setHoldToUnpauseDuration(value: Long) {
         preferenceManager.setLongValue(
@@ -35,10 +35,10 @@ class PauseSettingsScreenModel(context: Context) : ViewModel() {
         _holdToUnpauseDuration.postValue(value)
     }
 
-    fun setPauseTimeout(value: Long) {
+    fun setPauseTimeout(value: Int) {
         preferenceManager.setLongValue(
             PreferenceManager.Keys.PREFERENCE_KEY_PAUSE_TIMEOUT,
-            value
+            value.toLong()
         )
         _pauseTimeout.postValue(value)
     }
